Allow user to register functions from module script into the website using registerFunctionsInWindow method
This is a compromise of requiring the user to specify the script type as "module", which does not allow HTML elements or external JavaScript elements to access functions in module. This is especially useful for cases in which a button is created using sFront and the user wants the onclick attribute of the button to be defined in the same script as the button.